Add service layer

+++ app/src/main/java/com/ekotech/api/NetworkOptions.kt
@@ -3,6 +3,6 @@ package com.ekotech.api
 data class NetworkOptions(
     val apiKey: String,
     val hash: String,
-    val timeStamp: Int,
+    val timeStamp: Long,
     val offSet: Int
 )
